requirements.txt
annotated-types==0.6.0
anyio==4.3.0
audioread==3.0.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
deep-translator==1.11.4
distro==1.9.0
dtw-python==1.4.2
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.4
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
noisereduce==3.0.2
numba==0.59.1
numpy==1.26.4
openai==1.16.2
openai-whisper==20231117
packaging==24.0
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.13.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
sympy==1.12
threadpoolctl==3.4.0
tiktoken==0.6.0
torch==2.2.2
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
wheel==0.41.2
whisper==1.1.10
whisper-timestamped==1.15.3
google-generativeai
